requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
scipy==1.11.4
tqdm==4.66.4
pyyaml==6.0.1
rich==13.7.1
joblib==1.4.2
annoy==1.17.3
python-dateutil==2.9.0.post0
streamlit==1.36.0
plotly==5.23.0
